refactor(share): simplify preview-link code (/simplify pass)
Post-review cleanup over the preview-link + signing + rate-limit work:
- Extract duplicated base64url helpers to src/utils/base64url.ts; used by
  both privateUrlShare and publicShareLink (byte-identical before).
- Simplify gzip streamThrough to `new Response(stream).arrayBuffer()`; drop
  the hand-rolled collectStream reader loop.
- Remove dead buildPublicPreviewUrl (only its own tests used it — production
  uses encode -> sign -> assemble); tests now compose encodePublicDiagram +
  assemblePreviewUrl.
- og.ts: run the rate-limit round-trip only right before the render, after
  the cheap missing-param / bad-signature / decode rejects (junk requests no
  longer pay a WAF round-trip; every real render is still bounded).
- Drop the `export *` serverThemes shim from renderMermaid; renderAndPost
  imports ServerMermaidTheme from serverThemes directly.

No behavior change. api typecheck + build green; 135 tests pass.
